разделил код на модули и подключил точку входа
